Update SlidingUpPanelLayout.java
Add functionality for collapse on rotation and re-sizing main view on pane slide.

Added a new SlideState of MANUAL to describe a position not collapsed, expanded or anchored.

Created functionality to optionally automatically collapse the sliding pane on rotate. This is done by tracking the value of the current orientation. When the orientation changes onLayout is called. If the orientation has changed and the layout parameter has been set to true collapse the pane by setting mSlideOffset to 1.f and update the slide state. If mCollapseOnRotate is not true only update the value of the current orientation.

Created functionality to optionally automatically re-size the main view when the pane slides. The re-sizing of the main view is achieved during the execution of the draw method. If the parameter, mResizeMainView, is true the height of the main view is set to the local variable bottom. Bottom represents the space between the top of the sliding panel and the top of the draw-able area.

Added miscellaneous variables required to add this functionality. Added calls to setSlideState in collapsePane and expandPane to ensure the proper slide state is set.